This push is basically making averages in the same place
